requirements.txt
streamlit
yfinance
ta
google-generativeai
duckduckgo-search
python-dotenv
plotly